requirements.txt
boto>=2.49.0
boto3>=1.9.104
botocore>=1.12.104
bz2file>=0.98
certifi>=2018.11.29
chardet>=3.0.4
docutils>=0.14
elasticsearch>=7.0.0,<8.0.0
numpy>=1.16.2
gensim>=3.7.1
idna>=2.8
jmespath>=0.9.4
nltk>=3.4
google-api-python-client>=1.7.9
pandas>=0.24.1
pipenv>=2018.11.26
python-dateutil>=2.8.0
pytz>=2018.9
requests>=2.21.0
s3transfer>=0.2.0
scipy>=1.2.1
singledispatch>=3.4.0.3
six>=1.12.0
smart-open>=1.8.0
urllib3>=1.24.1
scikit-learn>=0.20
matplotlib>=3.0.3
python-rake>= 1.4.5



# Testing
requests>=2.21.0